Starting to clean up, I really should make a new branch though lol
