+++ firmware/include/state.h
@@ -1,0 +1,3 @@
+#ifndef STATE_H
+#define STATE_H
+#endif /* STATE_H */
